rename FileSystemHandler to FileSystemManager
